requirements.txt
altgraph==0.17.3
annotated-types==0.6.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blis==0.7.11
branca==0.6.0
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.3
confection==0.1.3
contourpy==1.0.7
curated-tokenizers==0.0.8
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.2/en_core_web_trf-3.7.2-py3-none-any.whl#sha256=dd56d80640411c2e81c3b0cfcb6068156a459c259ae11a4ea51c625c751e3e0a
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.4
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.39.4
fsspec==2023.9.2
future==0.18.3
geopandas==0.13.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud==0.34.0
google-cloud-documentai==2.20.1
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.10
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.1
openpyxl==3.1.2
osmnx==1.3.1.post0
outcome==1.2.0
overpy==0.6
packaging==23.1
pandas==1.5.2
parso==0.8.3
pdf2image==1.16.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
preshed==3.0.9
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pyparsing==3.0.9
PyPDF2==3.0.1
pyproj==3.5.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pywin32==306
pywin32-ctypes==0.2.0
pyzmq==25.1.0
regex==2023.10.3
requests==2.28.2
rsa==4.9
selenium==4.9.1
shapely==2.0.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.7.2
spacy-curated-transformers==0.2.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
tesseract==0.1.3
thinc==8.2.1
torch==2.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typer==0.9.0
typing_extensions==4.8.0
urllib3==1.26.15
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.3
webdriver-manager==3.8.6
wsproto==1.2.0
xlrd==2.0.1
